+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Bug Fixes
 - routing - fix rendering slow loading pages
   https://github.com/anvilistas/anvil-extras/pull/501
+- logging - add missing date format parameter
+  https://github.com/anvilistas/anvil-extras/pull/502
 
 ## Minor Changes
 - augment - use python implementation