+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   https://github.com/anvilistas/anvil-extras/pull/502
 
 ## Minor Changes
+- tabs - add better support for using faded colors e.g. `tabs.active_background = "#2196F344"`
+  https://github.com/anvilistas/anvil-extras/pull/483
 - augment - use python implementation
   https://github.com/anvilistas/anvil-extras/pull/488